+++ packages/jaeger-ui/src/components/TracePage/TracePageHeader.test.js
@@ -13,12 +13,9 @@
 // limitations under the License.
 
 import React from 'react';
-import sinon from 'sinon';
 import { shallow, mount } from 'enzyme';
 
 import TracePageHeader, { HEADER_ITEMS } from './TracePageHeader';
-import * as markers from './TracePageSearchBar.markers';
-import TracePageSearchBar from './TracePageSearchBar';
 
 describe('<TracePageHeader>', () => {
   const defaultProps = {
@@ -53,23 +50,5 @@ describe('<TracePageHeader>', () => {
       expect(item.contains(HEADER_ITEMS[i].title)).toBeTruthy();
       expect(item.contains(HEADER_ITEMS[i].renderer(defaultProps.trace))).toBeTruthy();
     });
-  });
-
-  it('calls updateTextFilter() function for onChange of the input', () => {
-    const updateTextFilter = sinon.spy();
-    const searchBarProps = {
-      updateTextFilter,
-      textFilter: defaultProps.textFilter,
-      prevResult: () => {},
-      nextResult: () => {},
-      resultCount: 0,
-    };
-    wrapper = shallow(<TracePageSearchBar {...searchBarProps} />);
-    const event = { target: { value: 'my new value' } };
-    wrapper
-      .find(`[data-test="${markers.IN_TRACE_SEARCH}"]`)
-      .first()
-      .simulate('change', event);
-    expect(updateTextFilter.calledWith('my new value')).toBeTruthy();
   });
 });
